E-0.0.30
- Collision Manager now is a sigleton
- Collision Manager circle math (WIP)
- Collision Manager (WIP)

+++ Engine/EngineDLL/BoundingCircle.cpp
@@ -1,6 +1,6 @@
 #include "BoundingCircle.h"
 
-BoundingCircle::BoundingCircle(glm::vec2 p, unsigned int r, bool setStatic, bool setTrigger) {
+BoundingCircle::BoundingCircle(glm::vec2 p, float r, bool setStatic, bool setTrigger) {
 	pos = p;
 	radius = r;
 	isStatic = setStatic;
